fix(gui-client/linux): bump `keyring` to 3.2.0 (#6345)
Closes #6175 (formerly closed as unable to replicate)

In some cases GNOME keyring will hide the attributes of credentials, so
`keyring` would throw an `Ambiguous` error when trying to sign in or
sign out, making it impossible to do anything.

The new version of `keyring` unlocks the keyring in this case, so the
attributes are shown and we can sign in correctly.

Thanks to @brotskydotcom for debugging and fixing this!