feat(connlib): remember recently connected gateways (#6361)
Previously, `connlib` would only send the currently connected gateways
to the portal upon a new connection intent. With our introduced idle
connection timeout, this could result in the portal choosing a different
gateway upon reconnecting to the resource.

To fix this, we introduce an LRU cache with at most 100 entries.
Iteration over the LRU cache happens in MRU order, meaning a recently
connected gateway will be at the front of the list.

We assume that this list is processed in order and thus still prefer
gateways that we are still connected to.

Related: #6347.