fix(android): Record crashlytics exception without crashing (#6475)
Instead of causing a NullPointerException, we should be more careful and
gracefully handle the case where the TunnelService's session pointer has
been cleared, but still record the exception in Crashlytics.

This patch also increases the robustness of the link properties DNS list
not having valid or useful DNS servers in them, and compacts the list
before sending to connlib, instead of potentially sending a JSON array
of nulls.

```
          Fatal Exception: java.lang.NullPointerException:
       at NetworkMonitor.onLinkPropertiesChanged(NetworkMonitor.kt:31)
       at android.net.ConnectivityManager$NetworkCallback.onAvailable(ConnectivityManager.java:4018)
       at android.net.ConnectivityManager$NetworkCallback.onAvailable(ConnectivityManager.java:3992)
       at android.net.ConnectivityManager$CallbackHandler.handleMessage(ConnectivityManager.java:4338)
       at android.os.Handler.dispatchMessage(Handler.java:106)
       at android.os.Looper.loopOnce(Looper.java:201)
       at android.os.Looper.loop(Looper.java:288)
       at android.os.HandlerThread.run(HandlerThread.java:67)
```


This fixes an issue seen by customer where the MDM solution is yanking
the TunnelService without graceful shutdown.