fix(android): Use DisconnectMonitor to shut tunnel down when VPN is disconnected (#6495)
When a user disconnects our VPN from system settings, Android does _not_
call any of the typical TunnelService callbacks. It does, however, send
a standard NetworkMonitor event we can intercept in the `onLost`
callback. We use that to shut down the TunnelService if we detect our
VPN network has been lost.

Unfortunately, network monitoring in Android is broad - we are notified
when _any_ network is lost, not just ours. So to filter others out, we
listen for `linkProperties` changing, and if the link addresses match
our tunnel IPv4 and IPv6, we save that network ID to match on in the
`onLost` callback later.

Resolves #5413