test(connlib): replace panics with ERROR logs (#6519)
In order to see all things that went wrong during a test run, we install
a special subscriber with the logger for `tunnel_test` that panics after
it has received at least 1 `ERROR` log.

The panics changed in this PR are still from a time when we didn't have
that. We change them so that we have better diagnostics for when they
get hit.