bug(rust/gui-client): if we can't raise the tunnel, sign out and delete the token.
Refs #6547

**Old behavior:** If anything goes wrong while connlib is starting, the IPC
service sends an error to the GUI, then the GUI silently implodes and we
can't export logs. When you restart the GUI, it reloads the token and implodes
again.

**New behavior:** The GUI doesn't silently implode, it silently signs out.
Which is still weird, but exporting logs will work.

This addresses an issue where a customer couldn't export logs while helping
us debug #6547.

```[tasklist]
- [ ] Manual test of behavior using fault injection or something
- [ ] Ready for review
```