fix(rust/gui-client): if we can't raise the tunnel, sign out (#6548)
Refs #6547 

Given a valid token is on disk, when the Client tries to auto-sign-in to
Firezone and something breaks during sign-in, then...

**Old behavior:** ...then the GUI silently implodes and we can't export
logs. When you restart the GUI, it reloads the token and immediately
implodes again.

**New behavior:** ...then the GUI doesn't silently implode, it silently
signs out. Which is still weird, but exporting logs will work.

This addresses an issue where a customer couldn't export logs while
helping us debug #6547.