chore(snownet): cache `connection` span (#6639)
Creating a new span is fairly expensive when it happens as part of a hot
function. Decapsulating packets in `snownet` is such a hot-function:


![image](https://github.com/user-attachments/assets/e1fb1672-7204-44e4-82f0-5471c000065a)

Previously, we created a new span for every packet that we decrypted
which accounted for ~3% of spent CPU time. We can optimise this and
remove some duplication by creating the span early and simply only
entering it every time we want it to be active. This results in
`boringtun`'s `decapsulate` being the most expensive function that
happens in `decapsulate`:


![image](https://github.com/user-attachments/assets/9050c798-d1eb-4d8c-93f0-64924125e5d9)