fix(client/windows): set MTU even if IPv6 is disabled
Refs #6547, this fixes a similar error message but it's not the same exact
issue.

When IPv6 is disabled on a system, our call to set the MTU was failing.
This patch allows some of the MTU-related syscalls to fail with a warning log.

To replicate the issue, run this command to set a registry value to disable
IPv6, then reboot the system:

`reg add "HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\Tcpip6\Parameters" /v DisabledComponents /t REG_DWORD /d 255 /f`